Finalização

+++ src/main/ecommerce/core/domain/service/DescontoService.java
@@ -1,6 +1,6 @@
 package main.ecommerce.core.domain.service;
 
-import main.ecommerce.core.domain.interfaces.DescontoRepository;
+import main.ecommerce.core.domain.contract.desconto.DescontoRepository;
 
 public class DescontoService {
     DescontoRepository descontoRepository;
@@ -9,7 +9,7 @@ public class DescontoService {
         this.descontoRepository = descontoRepository;
     }
 
-    public double aplicarDesconto(double total){
+    public double aplicarDesconto(float total){
         return this.descontoRepository.aplicarDesconto(total);
     }
 }
