+++ src/main/ecommerce/core/domain/service/PrimeiraCompra.java
@@ -1,11 +1,11 @@
 package main.ecommerce.core.domain.service;
 
-import main.ecommerce.core.domain.interfaces.DescontoRepository;
+import main.ecommerce.core.domain.contract.desconto.DescontoRepository;
 
 public class PrimeiraCompra implements DescontoRepository {
 
     @Override
-    public double aplicarDesconto(double total) {
-        return total - (total * (50.0 / 100.0));
+    public float aplicarDesconto(float total) {
+        return total - (total * (50 / 100));
     }
 }
